# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BookLib API",
    "version": "1.0.0",
    "description": "API for managing books, users, tags, comments, ratings, and plugins. Tags are always ensured from a resource file."
  },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "description": "Returns a list of all users. Requires JWT authentication.",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {"type": "integer"},
                      "username": {"type": "string"},
                      "email": {"type": "string"}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "tags": ["Users"],
        "summary": "User registration",
        "description": "Registers a new user with username, email, and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {"type": "string"},
                  "email": {"type": "string"},
                  "password": {"type": "string"}
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {"description": "User registered"},
          "400": {"description": "Missing fields"},
          "409": {"description": "User exists"}
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "User login",
        "description": "Authenticates a user and returns a JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {"type": "string"},
                  "password": {"type": "string"}
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "JWT token"},
          "401": {"description": "Invalid credentials"}
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a user by ID",
        "description": "Returns a user by their ID. Requires JWT authentication.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "integer"},
                    "username": {"type": "string"},
                    "email": {"type": "string"}
                  }
                }
              }
            }
          },
          "404": {"description": "User not found"}
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user by ID",
        "description": "Deletes a user by their ID. Requires JWT authentication.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {"description": "User deleted"},
          "404": {"description": "User not found"}
        }
      }
    },
    "/books": {
      "get": {
        "tags": ["Books"],
        "summary": "List books",
        "responses": {"200": {"description": "List of books"}}
      },
      "post": {
        "tags": ["Books"],
        "summary": "Add book",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "isbn": {"type": "string"},
                  "description": {"type": "string"},
                  "publish_year": {"type": ["integer", "string", "null"]},
                  "series": {"type": ["string", "null"]},
                  "tags": {"type": "array", "items": {"type": "string"}},
                  "plugin": {"type": "string", "description": "Plugin to use for enrichment (GoogleBooksPlugin, OpenLibraryPlugin)"}
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {"201": {"description": "Book added"}},
        "security": [{"BearerAuth": []}]
      }
    },
...
    "/tags": {
      "get": {
        "tags": ["Tags"],
        "summary": "Get all tags",
        "description": "Returns all tags in the database. Required tags are always loaded from a resource file.",
        "responses": {"200": {"description": "List of tags", "content": {"application/json": {"schema": {"type": "array", "items": {"type": "object", "properties": {"id": {"type": "integer"}, "name": {"type": "string"}}}}}}}}
      },
      "post": {
        "tags": ["Tags"],
        "summary": "Add a new tag",
        "description": "Adds a new tag to the database.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"name": {"type": "string"}}, "required": ["name"]}}}},
        "responses": {"201": {"description": "Tag added", "content": {"application/json": {"schema": {"type": "object", "properties": {"id": {"type": "integer"}, "name": {"type": "string"}}}}}}},
        "security": [{"BearerAuth": []}]
      }
    },
...
    "/books/{id}/comments": {
      "get": {
        "tags": ["Comments"],
        "summary": "Get comments for a book",
        "description": "Returns all comments for a book.",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {
            "description": "List of comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {"type": "integer"},
                      "text": {"type": "string"},
                      "user_id": {"type": "integer"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Comments"],
        "summary": "Add a comment for a book",
        "description": "Adds a comment for a book. Requires JWT authentication.",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {"type": "string"}
                },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "integer"}
                  }
                }
              }
            }
          },
          "400": {"description": "Missing text"}
        },
        "security": [{"BearerAuth": []}]
      }
    },
...
    "/books/{id}/ratings": {
      "get": {
        "tags": ["Ratings"],
        "summary": "Get average rating for a book",
        "description": "Returns the average rating for a book.",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {
            "description": "Average rating",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "average": {"type": ["number", "null"]}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Ratings"],
        "summary": "Add a rating for a book",
        "description": "Adds a rating for a book. Requires JWT authentication. Each user can only rate a book once.",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {"type": "integer"}
                },
                "required": ["rating"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rating added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "integer"}
                  }
                }
              }
            }
          },
          "400": {"description": "Missing rating"},
          "409": {"description": "User has already rated this book"}
        },
        "security": [{"BearerAuth": []}]
      }
    },
...
    "/plugins/load": {"post": {"summary": "Load plugin", "requestBody": {"required": true}, "responses": {"200": {"description": "Plugin loaded"}}}},
    "/plugins/unload": {"post": {"summary": "Unload plugin", "requestBody": {"required": true}, "responses": {"200": {"description": "Plugin unloaded"}}}},
...
  }
}